requirements.txt

Django>=4.2
djangorestframework
psycopg2-binary
redis
celery
django-redis
faker
python-dotenv
